Jelly: Convert to ViewModel + flows + ListAdapter

+++ app/src/main/java/org/lineageos/jelly/flow/FavoriteFlow.kt
@@ -37,7 +37,7 @@ class FavoriteFlow(private val context: Context) : QueryFlow<Favorite> {
     ) { it, indexCache ->
         var i = 0
         Favorite(
-            it.getInt(indexCache[i++]),
+            it.getLong(indexCache[i++]),
             it.getString(indexCache[i++]),
             it.getString(indexCache[i++]),
             it.getInt(indexCache[i++])
